requirements.txt
fastapi
uvicorn
langchain
openai
pypdf
chromadb
streamlit
python-dotenv
requests
